+++ models/locations.go
@@ -17,5 +17,4 @@ type Location struct {
 
 	Instance Instance `bun:"rel:has-one,join:instance_id=id"`
 	Marker   Marker   `bun:"rel:has-one,join:marker_id=code"`
-	Blocks   []Block  `bun:"rel:has-many,join:id=location_id"`
 }