+++ repositories/share_link_repository.go
@@ -68,7 +68,6 @@ func (r *ShareLinkRepository) GetByID(ctx context.Context, id string) (*models.S
 		Relation("Template").
 		Relation("Template.Settings").
 		Relation("Template.Locations").
-		Relation("Template.Locations.Blocks").
 		Relation("Template.Locations.Marker").
 		Scan(ctx)
 	return link, err