+++ docs/changelog.md
@@ -45,6 +45,7 @@ The most exciting change is the ability to use blocks in multiple contexts, not 
 ### Removed
 
 - **Breaking change:** `ClueRepository`, `ClueService`, and all clue-related views and logic. 
+- **Breaking change:** `Locations.Blocks` relation. Only used in tests, so switched to service calls instead.
   - Clues are now automatically migrated to [Random Clue Blocks](/docs/user/blocks/random-clue) when the migration is run.
 
 [Full Changelog](https://github.com/nathanhollows/Rapua/releases/tag/v4.4.0)